requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
gunicorn>=21.2.0
stripe>=7.0.0
psycopg2-binary>=2.9.0
pdfplumber>=0.10.0
python-docx>=1.0.0
PyPDF2>=3.0.0
